Fix several windows build issues (#309)
The mingw build has rotten somewhat, there are some type errors
and data access errors.
The fix is not yet complete, but it's a step forward

+++ src/anyp/Uri.cc
@@ -657,7 +657,7 @@ urlMakeAbsolute(const HttpRequest * req, const char *relUrl)
 }
 
 int
-matchDomainName(const char *h, const char *d, uint flags)
+matchDomainName(const char *h, const char *d, uint8_t flags)
 {
     int dl;
     int hl;
